Real or Not? NLP with Disaster Tweets
Twitter has become an important communication channel in times of emergency. The ubiquitousness of smartphones enables people to announce an emergency they’re observing in real-time. Because of this, more agencies are interested in programatically monitoring Twitter (i.e. disaster relief organizations and news agencies).
This particular challenge is perfect for data scientists looking to get started with Natural Language Processing